Test: Move `case_notification` in integration s.
